+++ src/IgorAux/src/iaString.h
@@ -56,7 +56,7 @@ namespace IgorAux
 
 		/*! this defines a position that is out of range
 		*/
-		static const size_t INVALID_POSITION = 0xffffffffffffffff;
+		static const uint64 INVALID_POSITION = 0xffffffffffffffff;
 
 		/*! default ctor does nothing
 		*/
@@ -67,14 +67,14 @@ namespace IgorAux
 		\param text multipbyte string
 		\param lenght if input string is not zero terminated or a custom lenght is needed
 		*/
-		iaString(const char* text, const size_t lenght = INVALID_POSITION);
+		iaString(const char* text, const uint64 lenght = INVALID_POSITION);
 
 		/*! copies unicode string
 
 		\param text unicode string
 		\param lenght if input string is not zero terminated or a custom lenght is needed
 		*/
-		iaString(const wchar_t* text, const size_t lenght = INVALID_POSITION);
+		iaString(const wchar_t* text, const uint64 lenght = INVALID_POSITION);
 
         /*! creates a string with only one character
 
@@ -201,19 +201,19 @@ namespace IgorAux
 		/*! \returns charachter at given position for manipulation
 		\param index the position to get the character from
 		*/
-		wchar_t& operator[](const size_t index);
+		wchar_t& operator[](const uint64 index);
 
 		/*! \returns charachter at given position
 		\param index the position to get the character from
 		*/
-		const wchar_t& operator[](const size_t index) const;
+		const wchar_t& operator[](const uint64 index) const;
 
 		/*! \returns specified substring
 
 		\param pos substring from
 		\param len substring lenght
 		*/
-		iaString getSubString(const size_t pos, const size_t len = INVALID_POSITION) const;
+		iaString getSubString(const uint64 pos, const uint64 len = INVALID_POSITION) const;
 
         /*! plsits a string based on delimiter in to tokens (or substrings)
 
@@ -243,7 +243,7 @@ namespace IgorAux
 		\param characters the characters to search for
 		\param from optional parameter to define start index to search from
 		*/
-		size_t findFirstOf(const wchar_t* characters, const size_t from = INVALID_POSITION) const;
+		uint64 findFirstOf(const wchar_t* characters, const uint64 from = INVALID_POSITION) const;
 
         /*! \returns position of first occurence of specified character
 
@@ -252,7 +252,7 @@ namespace IgorAux
         \param character the character to search for
         \param from optional parameter to define start index to search from
         */
-        size_t findFirstOf(const wchar_t character, const size_t from = INVALID_POSITION) const;
+        uint64 findFirstOf(const wchar_t character, const uint64 from = INVALID_POSITION) const;
 
 		/*! \returns position of first occurence NOT of specified characters
 
@@ -261,7 +261,7 @@ namespace IgorAux
 		\param characters the characters to NOT search for
 		\param from optional parameter to define start index to search from
 		*/
-		size_t findFirstNotOf(const wchar_t* characters, const size_t from = INVALID_POSITION) const;
+		uint64 findFirstNotOf(const wchar_t* characters, const uint64 from = INVALID_POSITION) const;
 
         /*! \returns position of first occurence NOT of specified character
 
@@ -270,7 +270,7 @@ namespace IgorAux
         \param character the character to NOT search for
         \param from optional parameter to define start index to search from
         */
-        size_t findFirstNotOf(const wchar_t character, const size_t from = INVALID_POSITION) const;
+        uint64 findFirstNotOf(const wchar_t character, const uint64 from = INVALID_POSITION) const;
 
 		/*! \returns position of last occurence of specified character
 
@@ -278,7 +278,7 @@ namespace IgorAux
 
 		\param character the character to search for
 		*/
-		size_t findLastOf(const wchar_t character) const;
+		uint64 findLastOf(const wchar_t character) const;
 
         /*! \returns position of last occurence of specified characters
 
@@ -286,7 +286,7 @@ namespace IgorAux
 
         \param characters the characters to search for
         */
-        size_t findLastOf(const wchar_t* characters) const;
+        uint64 findLastOf(const wchar_t* characters) const;
 
 		/*! empties the string
 		*/
@@ -298,7 +298,7 @@ namespace IgorAux
 
 		/*! \returns the size (aka character count) of the string without ending zero
 		*/
-		size_t getSize() const;
+		uint64 getSize() const;
 
 		/*! \returns pointer to raw data
 		*/
@@ -310,7 +310,7 @@ namespace IgorAux
 		\param size size of preallocated buffer
 		\returns actual size of written data including ending zero
 		*/
-		size_t getData(char* buffer, const size_t size) const;
+		uint64 getData(char* buffer, const uint64 size) const;
 
         /*! returns utf8 encoded string
 
@@ -318,20 +318,20 @@ namespace IgorAux
         \param size the size of the destination buffer
         \returns bytes written to destination buffer
         */
-        size_t getUTF8(char* buffer, const size_t size) const;
+        uint64 getUTF8(char* buffer, const uint64 size) const;
 
         /*! calclates the size in bytes of the string encoded in utf8
 
         \returns size in bytes
         */
-        size_t getUTF8Size() const;
+        uint64 getUTF8Size() const;
 
         /*! set string with utf8 encoded string
 
         \param buffer source buffer with utf8 string
         \param size the size of the source buffer
         */
-        void setUTF8(const char* buffer, const size_t size);
+        void setUTF8(const char* buffer, const uint64 size);
 
         /*! reverses the string
         */
@@ -365,7 +365,7 @@ namespace IgorAux
 
         /*! string lenght without ending zero
         */
-        size_t _charCount = 0;
+        uint64 _charCount = 0;
 
         /*! pointer to actual data
         */
@@ -376,14 +376,14 @@ namespace IgorAux
 		\param text the data to set
         \param size the size of the buffer
 		*/
-		void setData(const wchar_t* text, const size_t size = INVALID_POSITION);
+		void setData(const wchar_t* text, const uint64 size = INVALID_POSITION);
 
 		/*! converts multibyte string to unicode string
 
 		\param text the data to set
 		\param size the size of the buffer
 		*/
-		void setData(const char* text, const size_t size = INVALID_POSITION);
+		void setData(const char* text, const uint64 size = INVALID_POSITION);
 
         static iaString intToStrInternal(int32 x, int d);
